+++ src/main/java/insane96mcp/iguanatweaksexpanded/data/generator/ITEGlobalLootModifierProvider.java
@@ -3,7 +3,6 @@ package insane96mcp.iguanatweaksexpanded.data.generator;
 import insane96mcp.iguanatweaksexpanded.module.experience.enchanting.EnchantingFeature;
 import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
 import insane96mcp.iguanatweaksexpanded.module.mining.SoulSteel;
-import insane96mcp.iguanatweaksexpanded.module.movement.minecarts.Minecarts;
 import insane96mcp.iguanatweaksexpanded.module.world.coalfire.CoalCharcoal;
 import net.minecraft.data.PackOutput;
 import net.minecraftforge.common.data.GlobalLootModifierProvider;
@@ -13,10 +12,8 @@ public class ITEGlobalLootModifierProvider extends GlobalLootModifierProvider {
         super(output, modid);
     }
 
-
     @Override
     protected void start() {
-        Minecarts.addGlobalLoot(this);
         CoalCharcoal.addGlobalLoot(this);
         EnchantingFeature.addGlobalLoot(this);
         NewEnchantmentsFeature.addGlobalLoot(this);